Revert "refactor(kitchenpos): DisplayedName"
This reverts commit 2b316682a0db4877d5d3360d6ee89f39f7a8fce9.
