+++ src/main/java/kitchenpos/products/infra/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.domain;
+package kitchenpos.products.infra;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
