+++ src/test/java/kitchenpos/products/application/FakePurgomalumClient.java
@@ -1,6 +1,6 @@
-package kitchenpos.common.infra;
+package kitchenpos.products.application;
 
-import kitchenpos.common.domain.PurgomalumClient;
+import kitchenpos.products.infra.PurgomalumClient;
 
 import java.util.Arrays;
 import java.util.List;
